feat(application) : refactor DTOs and Validators for Improved Modularity
- Restructure Data Transfer Objects (DTOs) into specific namespaces: `CoreXCrud.DTOs.OrderDetailDtos`, `CoreXCrud.DTOs.OrderDtos`, `CoreXCrud.DTOs.ProductDtos`, and `CoreXCrud.DTOs.UserDtos` for better clarity and maintainability.
- Update the `MappingProfile` class to include new DTO namespaces and ensure correct mappings between models and DTOs.
- Revise validators (`OrderDetailValidator`, `OrderValidator`, `ProductValidator`, and `UserValidator`) to reference the new namespaces.
- Modify `Program.cs` to incorporate the updated mappings and validators.
- Add migration files (`20250610234634_CreateEntities.cs` and `20250610234634_CreateEntities.Designer.cs`) to set up the database schema for new entities.
- Update `appsettings.json` to change the database connection string, reflecting a potential change in the database server.
- These changes enhance the organization and modularity of the codebase, making it easier to manage and extend.

<-- This commit message was generated with GitHub Copilot in Visual Studio. -->
