+++ Controllers/OrdersController.cs
@@ -1,11 +1,11 @@
 ﻿using AutoMapper;
-using CoreXCrud.DTOs;
 using CoreXCrud.Models;
 using CoreXCrud.Repositories;
 using FluentValidation;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Authorization;
 using Serilog;
+using CoreXCrud.DTOs.OrderDtos;
 
 namespace CoreXCrud.Controllers
 {
