+++ appsettings.json
@@ -7,7 +7,7 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DefaultConnection": "Server=PINARIISKO;Database=CoreXCrudDB;Trusted_Connection=True;TrustServerCertificate=True;MultipleActiveResultSets=true"
+    "DefaultConnection": "Server=YASINRAMAZANGOK;Database=CoreXCrudDB;Trusted_Connection=True;TrustServerCertificate=True;MultipleActiveResultSets=true"
   },
   "AppSettings": {
     "ProjectName": "",
